refactor: rename project to Cartify
