+++ docker/docker-compose.yml
@@ -1,49 +1,49 @@
 services:
   postgres:
     image: postgres:15-alpine
-    container_name: ecommerce_postgres
+    container_name: cartify_postgres
     environment:
-      POSTGRES_DB: ecommerce
+      POSTGRES_DB: cartify
       POSTGRES_USER: user
       POSTGRES_PASSWORD: password
     ports:
       - "5432:5432"
     networks:
-      - ecommerce-net
+      - cartify-net
   
-  ecommerce:
+  cartify:
     build: ../
-    container_name: ecommerce_app
+    container_name: cartify_app
     ports:
       - "8080:8080"
     depends_on:
       - postgres
     networks:
-      - ecommerce-net
+      - cartify-net
     environment:
-      SPRING_DATASOURCE_URL: jdbc:postgresql://postgres:5432/ecommerce
+      SPRING_DATASOURCE_URL: jdbc:postgresql://postgres:5432/cartify
       SPRING_DATASOURCE_USERNAME: user
       SPRING_DATASOURCE_PASSWORD: password
 
   redis:
     image: redis:alpine
-    container_name: ecommerce_redis
+    container_name: cartify_redis
     ports:
       - "6379:6379"
     networks:
-      - ecommerce-net
+      - cartify-net
 
   zookeeper:
     image: wurstmeister/zookeeper
-    container_name: ecommerce_zookeeper
+    container_name: cartify_zookeeper
     ports:
       - "2181:2181"
     networks:
-      - ecommerce-net
+      - cartify-net
 
   kafka:
     image: wurstmeister/kafka
-    container_name: ecommerce_kafka
+    container_name: cartify_kafka
     ports:
       - "9092:9092"
     environment:
@@ -53,7 +53,7 @@ services:
     depends_on:
       - zookeeper
     networks:
-      - ecommerce-net
+      - cartify-net
 
 networks:
-  ecommerce-net:
+  cartify-net:
