+++ pom.xml
@@ -9,9 +9,9 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
-	<artifactId>e-commerce</artifactId>
+	<artifactId>cartify</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>e-commerce</name>
+	<name>cartify</name>
 	<description>Demo project for Spring Boot</description>
 	<url/>
 	<licenses>
