+++ src/main/java/com/example/cartify/auth/AuthApplicationService.java
@@ -1,0 +1,5 @@
+package com.example.cartify.auth;
+
+public class AuthApplicationService {
+    
+}
