+++ src/main/java/com/example/cartify/auth/dto/AuthResponse.java
@@ -1,0 +1,16 @@
+package com.example.cartify.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class AuthResponse {
+    private String token;
+}
