+++ src/main/java/com/example/cartify/auth/dto/LoginRequest.java
@@ -1,0 +1,13 @@
+package com.example.cartify.auth.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class LoginRequest {
+    private String email;
+    private String password;
+}
