+++ src/main/java/com/example/cartify/auth/dto/RegisterRequest.java
@@ -1,0 +1,15 @@
+package com.example.cartify.auth.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class RegisterRequest {
+    private String firstname;
+    private String lastname;
+    private String email;
+    private String password;
+}
