+++ src/main/java/com/example/cartify/auth/model/Role.java
@@ -1,0 +1,6 @@
+package com.example.cartify.auth.model;
+
+public enum Role {
+    USER,
+    ADMIN
+}
