+++ src/main/java/com/example/cartify/auth/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.example.cartify.auth.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.cartify.auth.model.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
