+++ src/main/java/com/example/cartify/cartifyApplication.java
@@ -1,0 +1,15 @@
+package com.example.cartify;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+
+@EnableMethodSecurity
+@SpringBootApplication
+public class cartifyApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(cartifyApplication.class, args);
+	}
+
+}
