+++ src/main/java/com/example/cartify/config/OpenApiConfig.java
@@ -1,0 +1,32 @@
+package com.example.cartify.config;
+
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.security.*;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.List;
+
+@Configuration
+public class OpenApiConfig {
+
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+            .info(new Info()
+                .title("Cartify API")
+                .version("1.0")
+                .description("API documentation for Cartify")
+            )
+            .components(new Components()
+                .addSecuritySchemes("BearerAuth",
+                    new SecurityScheme()
+                        .type(SecurityScheme.Type.HTTP)
+                        .scheme("bearer")
+                        .bearerFormat("JWT")
+                ))
+            .addSecurityItem(new SecurityRequirement().addList("BearerAuth", List.of("read", "write")));
+    }
+}
