+++ src/main/java/com/example/cartify/product/dto/ProductRequest.java
@@ -1,0 +1,17 @@
+package com.example.cartify.product.dto;
+
+import lombok.*;
+
+import java.math.BigDecimal;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ProductRequest {
+    private String name;
+    private String description;
+    private BigDecimal price;
+    private String imageUrl;
+}
