+++ src/main/java/com/example/cartify/product/dto/ProductResponse.java
@@ -1,0 +1,22 @@
+package com.example.cartify.product.dto;
+
+import java.math.BigDecimal;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ProductResponse {
+    private Long id;
+    private String name;
+    private String description;
+    private BigDecimal price;
+    private String imageUrl;
+}
