--- src/main/java/com/example/ecommerce/ECommerceApplication.java
@@ -1,15 +1,0 @@
-package com.example.ecommerce;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-
-@EnableMethodSecurity
-@SpringBootApplication
-public class ECommerceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ECommerceApplication.class, args);
-	}
-
-}
