--- src/main/java/com/example/ecommerce/auth/AuthApplicationService.java
@@ -1,5 +1,0 @@
-package com.example.ecommerce.auth;
-
-public class AuthApplicationService {
-    
-}
