--- src/main/java/com/example/ecommerce/auth/controller/AuthController.java
@@ -1,32 +1,0 @@
-package com.example.ecommerce.auth.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.example.ecommerce.auth.dto.AuthResponse;
-import com.example.ecommerce.auth.dto.LoginRequest;
-import com.example.ecommerce.auth.dto.RegisterRequest;
-import com.example.ecommerce.auth.service.AuthenticationService;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequestMapping("/api/v1/auth")
-@RequiredArgsConstructor
-public class AuthController {
-    
-    private final AuthenticationService authService;
-
-    @PostMapping("/register")
-    public ResponseEntity<AuthResponse> register(@RequestBody RegisterRequest request) {
-        return ResponseEntity.ok(authService.register(request));
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<AuthResponse> login(@RequestBody LoginRequest request) {
-        return ResponseEntity.ok(authService.login(request));
-    }
-}
