--- src/main/java/com/example/ecommerce/auth/dto/AuthResponse.java
@@ -1,16 +1,0 @@
-package com.example.ecommerce.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class AuthResponse {
-    private String token;
-}
