--- src/main/java/com/example/ecommerce/auth/dto/LoginRequest.java
@@ -1,13 +1,0 @@
-package com.example.ecommerce.auth.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class LoginRequest {
-    private String email;
-    private String password;
-}
