--- src/main/java/com/example/ecommerce/auth/dto/RegisterRequest.java
@@ -1,15 +1,0 @@
-package com.example.ecommerce.auth.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class RegisterRequest {
-    private String firstname;
-    private String lastname;
-    private String email;
-    private String password;
-}
