--- src/main/java/com/example/ecommerce/auth/model/Role.java
@@ -1,6 +1,0 @@
-package com.example.ecommerce.auth.model;
-
-public enum Role {
-    USER,
-    ADMIN
-}
