--- src/main/java/com/example/ecommerce/auth/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.example.ecommerce.auth.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.ecommerce.auth.model.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
