--- src/main/java/com/example/ecommerce/auth/service/AuthenticationService.java
@@ -1,50 +1,0 @@
-package com.example.ecommerce.auth.service;
-
-import com.example.ecommerce.auth.model.Role;
-import com.example.ecommerce.auth.model.User;
-
-import org.springframework.security.authentication.*;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-import com.example.ecommerce.auth.dto.*;
-import com.example.ecommerce.auth.repository.UserRepository;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class AuthenticationService {
-
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-    private final JwtService jwtService;
-    private final AuthenticationManager authenticationManager;
-
-    public AuthResponse register(RegisterRequest request) {
-        var user = User.builder()
-                .firstname(request.getFirstname())
-                .lastname(request.getLastname())
-                .email(request.getEmail())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .role(Role.USER)
-                .build();
-        
-        userRepository.save(user);
-        var token = jwtService.generateToken(user.getEmail());
-        return AuthResponse.builder().token(token).build();
-    }
-
-    public AuthResponse login(LoginRequest request) {
-        authenticationManager.authenticate(
-            new UsernamePasswordAuthenticationToken(request.getEmail(), request.getPassword())
-        );
-
-        var user = userRepository.findByEmail(request.getEmail())
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-
-        var token = jwtService.generateToken(user.getEmail());
-        return AuthResponse.builder().token(token).build();
-    }
-}
