--- src/main/java/com/example/ecommerce/product/dto/ProductRequest.java
@@ -1,17 +1,0 @@
-package com.example.ecommerce.product.dto;
-
-import lombok.*;
-
-import java.math.BigDecimal;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ProductRequest {
-    private String name;
-    private String description;
-    private BigDecimal price;
-    private String imageUrl;
-}
