--- src/main/java/com/example/ecommerce/product/dto/ProductResponse.java
@@ -1,22 +1,0 @@
-package com.example.ecommerce.product.dto;
-
-import java.math.BigDecimal;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ProductResponse {
-    private Long id;
-    private String name;
-    private String description;
-    private BigDecimal price;
-    private String imageUrl;
-}
