--- src/main/java/com/example/ecommerce/product/service/ProductService.java
@@ -1,72 +1,0 @@
-package com.example.ecommerce.product.service;
-
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-
-import com.example.ecommerce.product.dto.ProductRequest;
-import com.example.ecommerce.product.dto.ProductResponse;
-import com.example.ecommerce.product.model.Product;
-import com.example.ecommerce.product.repository.ProductRepository;
-
-import lombok.RequiredArgsConstructor;
-
-
-@Service
-@RequiredArgsConstructor
-public class ProductService {
-    private final ProductRepository productRepository;
-
-    public ProductResponse createProduct(ProductRequest request) {
-        Product product = Product.builder()
-                .name(request.getName())
-                .description(request.getDescription())
-                .price(request.getPrice())
-                .imageUrl(request.getImageUrl())
-                .build();
-        
-        product = productRepository.save(product);
-        return toResponse(product);
-    }
-
-    public List<ProductResponse> getAllProducts() {
-        return productRepository.findAll()
-                .stream()
-                .map(this::toResponse)
-                .toList();
-    }
-
-    public ProductResponse getProductById(Long id) {
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Product Not Found"));
-        
-        return toResponse(product);
-    }
-
-    public ProductResponse updateProduct(Long id, ProductRequest request) {
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Product Not Found"));
-
-        product.setName(request.getName());
-        product.setDescription(request.getDescription());
-        product.setPrice(request.getPrice());
-        product.setImageUrl(request.getImageUrl());
-                
-        return toResponse(productRepository.save(product));
-    }
-
-    public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
-    }
-
-    private ProductResponse toResponse(Product product) {
-        return ProductResponse.builder()
-                .id(product.getId())
-                .name(product.getName())
-                .description(product.getDescription())
-                .price(product.getPrice())
-                .imageUrl(product.getImageUrl())
-                .build();
-
-    }
-}
