+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   datasource:
-    url: jdbc:postgresql://postgres:5432/ecommerce
+    url: jdbc:postgresql://postgres:5432/cartify
     username: user
     password: password
   jpa:
@@ -10,7 +10,7 @@ spring:
     database-platform: org.hibernate.dialect.PostgreSQLDialect
 
   redis:
-    host: ecommerce_redis
+    host: cartify_redis
     port: 6379
 
   kafka:
