+++ src/test/java/com/example/cartify/CartifyApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.cartify;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class cartifyApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
