--- src/test/java/com/example/ecommerce/ECommerceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.ecommerce;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ECommerceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
